Add support for embedding PNG images
The new \image markup command embeds an image in PNG or EPS format,
based on the file extension.

The purpose of not making a "\pngfile" command is to make it more
natural in the future to add support for JPEG and SVG images.  It will
be easier for the user not to care about the file format and not to have
the documentation for embedding images scattered over 4 places in the
list of markup commands.

The \image command adds a white background by default.  The reason is
that in the PS backend, it is not possible to support transparency for
PNG images.  We want the default to be consistent across backends in
order not to add a hurdle for users when we switch to the Cairo backend.
The background-color property can be set to #f in order to disable the
background, which only works in the SVG and Cairo backends.

\epsfile is kept for now because it will have to be eventually
deprecated at least for PDF/PNG output due to the switch to Cairo,
meaning that most \epsfile invocations will have to be converted anyway.
Having to convert things just once is easier for users than doing it
twice (\epsfile → \image ... "x.eps" → \image ... "x.png").

This commit only has the Cairo backend implementation, but later commits
add it to the default backends as well.

The Cairo implementation is quite simple: read the PNG image into an
image surface (Cairo internally uses LibPNG for this), and paint this
surface onto the output surface.