binaries: Add Brewfile for macOS
While there is a homebrew module for Ansible, it comes with the caveat
that "homebrew must already be installed on the target system" and
implementing an automatic installation via a git-clone does not work
easily because Homebrew checks out a stable branch without a remote
equivalent. As we only need to install a small number of packages
via Homebrew on macOS, it is instead easier to just maintain a Brewfile.