Refactor colors
Add a normalize-color Scheme function that takes a color in any style
(color list, CSS) and returns an (r g b a) list.  Reimplement
ly:stencil-in-color and stencil-with-color in terms of normalize-color.

Having normalize-color will enable background handling for PNG images in
the PS backend, where the 'pngfile stencil expression will need to
include a color.

The code is moved to Scheme because I got tired of doing complicated
type checks in C++.