Do not disturb printing compound bar lines if span bar is false
Previously a user defined bar line with span bar #f like
\defineBarLine "...;|" #'(#t #t #f)
resulted in only printing the first character of the bar-type.
I.e above resulted in a single thick bar line.
Now we provide a list with #f-values of same length as the relevant
list for bar line characters, instead of the previous (list "").
Warning and fall back is added if span bar is neither string nor
boolean.
A regtest is added as well.

Closes #6521