Removed the "arabic" note language
- The "arabic" note language was only necessary for writing 5/2 and 7/2
alterations, something that can't be achieved with the standard note
languages. This feature is not required anymore as stated by the author:
https://lists.gnu.org/archive/html/lilypond-devel/2023-01/msg00176.html
The language was also not very well designed, since it uses English
note names buit with italian/catalan-like alterations and
double-sharps/double-flats could not be written anymore.
- Removed the documentation part that described that language
- Removed the references to hel-arabic.ly in the documentation
where it is not used
- Removed the glpyh list from hel-arabic.ly because without the 5/2 and 7/2
intervals it can use the one included in arabic.ly