musicxml2ly: remove the option to parse with lxml
This option was not covered by routine regression testing, so it is not
surprising that it failed when it was recently tested manually.

On a large score, lxml was seen to save seconds of run time compared to
minidom.  The feeling is that that isn't worth the maintenance cost.

musicxml2ly --lxml now prints a warning and continues using minidom.