Documentation: Use section names for buttons
For now, the differences are minimal because most node's (that were
used so far) match the section names. The most frequent change is
that we get the escaped &rsquo; instead of an ASCII apostrophe, and
proper rendering of some Texinfo commands such as code. However, it
matches the spirit of xrefautomaticsectiontitle, which is the only
mode in texi2html and the default in the HTML output of texi2any.