Volta_engraver cleaning
Remove the code to merge volta numbers from different layers of repeat
structure.  This is cruft from the early days of \volta which has been
unneeded since the engraver was enhanced to create nested brackets.

A noteworthy enhancement in this commit is that VoltaBracket and
VoltaBracketSpanner grobs are now created with the stream events that
caused them (if stream events caused them).  This is necessary but not
sufficient to support adjusting them with \tweak.