LM: Fix indexing of `\\`
In PDF output, `\\` was previously shown in the index as if it would have
started with letter `Λ` (uppercase Greek Lambda).  The reason for this
bizarre result was that we use character code 3 as a sort-key postfix in
`common-macros.itexi`.  Because `\` gets removed for sorting purposes (we
want `\foo` to appear next to 'foo' in the index) character code 3 remained
visible in the output of the `texindex` sorting program, and the roman fonts
of Computer Modern have `Λ` at exactly this glyph index position.

This commit adds a `@sortas` key with a substitution for `\` that actually
survives the removal of `\` characters in TeX output.