doc: Prepare for future `texinfo.tex` versions 1/2
In the beginning of February 2023 the implementation of the `@XXindex`
commands in the git version of `texinfo.tex` changed: they no longer worked
in the middle of a paragraph (except if immediately preceded by `@item`);
instead, they caused a paragraph break.  This change was (temporarily?)
reverted in the end of March 2023 to avoid backward incompatibilities.

Note that the change didn't come out of the blue; there are good reasons
(mainly for HTML output) to make `@XXindex` end a paragraph.  For this
reason it makes sense to update the LilyPond documentation, avoiding
inter-paragraph indexing.

The following command was used to identify problematic index lines.

```
pcregrep \
  --exclude-dir=.git \
  --multiline \
  --recursive \
  '(?x)
     ^ (?! @c[ ] | @item | @cindex | @anchor | @funindex ) .* \S .* $ \s
     ( ^ ( @c[ ] | @cindex | @anchor | @funindex ) .* $ \s )*
     ( ^ ( @cindex | @funindex ) .* $ \s )+
     ( ^ ( @c[ ] | @cindex | @anchor | @funindex ) .* $ \s )*
     ^ (?! @c[ ] | @cindex | @anchor | @funindex ) .* \S
  ' \
  .
```

This commit moves all indexing commands to the beginning of `@predefined`
blocks, followed by an empty line.  It also removes redundant index entries
in the German translation.