-dembed-source-code: Use base name of embedded file only
Absolute paths aren't meaningful to embed (since the receiver can't save
with the same path on their computer), and they lead to sharing
information about the path where the PDF was generated, which might be
unwanted. Even relative paths (with at least one directory separator,
that is) aren't handled consistently by PDF viewers. This change does
mean that with \include "directory/included.ly", the output PDF file
does not preserve the information that included.ly is in a subfolder,
and it does also introduce a conflict between \include
"directory1/included.ily" and \include "directory2/included.ily" (with a
warning), but this is the price to pay for unsurprising and consistent
handling of the attachments.

Fixes #6526