Add VoltaBracket.musical-length to simplify overhang
Shortening volta brackets with the context property voltaSpannerDuration
has some drawbacks that this new grob property overcomes.

To shorten a bracket with voltaSpannerDuration, the engraver reads it in
every timestep and if it finds the right value at the right time, ends
the bracket there.  While the user might be able to set the property
ahead of time and leave it, the user must still write the music in a way
that the engraver visits the desired ending timestep.

VoltaBracket.musical-length is read when the bracket is created, and the
engraver can guarantee that the ending timestep is visited.

Controlling bracket length with a grob property also opens a door to
future \tweak support.

The former volta-overhang test, which uses manual repeat commands, has
been renamed to volta-overhang-manual.