Stop separating alternatives with \bar ""
One of the changes supporting VoltaBracket.musical-length allows a
VoltaBracket to end at the current command column when there is no
BarLine.

Now that no BarLine is required, we can remove volta-span handling from
Bar_engraver.

One good effect can be seen in regression test
repeat-bar-number-engraver-order.  A volta bracket ending in the middle
of a measure used to cause a parenthesized bar number to appear there
because of the invisible bar line.  Now there is no bar line, so there
is no bar number.  (On the other hand, maybe an invisible bar line
should not have caused a bar number anyway.)

Another good effect can be seen in regression test
repeat-volta-body-grace, where it is clearer that the alternative begins
after the grace note.  The test is silly, but the new output is still a
clearer depiction of the ly code.