Rename 'feta font-family value to 'music
This reflects the fact that Feta is not the only possible music font;
the user can choose a different one.

Note that this name has been internally used for a long time, but the
renaming is more pressing now that the user interface for font selection
makes it appear:

\paper {
  fonts.feta = "..."
}

becomes

\paper {
  fonts.music = "..."
}

with this commit.