Remove support for building with a different internal unit
For fun, I tried to build LilyPond with CXXFLAGS="-DPOINTS". The result
of { c' } was a score with a too small staff, too large font glyphs and
a tagline flowing out of the margins. Building LilyPond with points as
internal units does not have a clear use case and the ship seems to have
sailed.