parmesan-noteheads.mf (draw_mensural_longa): Fix numerical instability
In the previous code, METAFONT didn't find an intersection between the
notehead of some mensural glyphs and the stem in some cases, causing either
an incorrect attachment point or weird shapes (without any warnings due to
the way the outline was constructed).

Note that METAPOST (and thus the generation of the Emmentaler OpenType font)
was not affected, by pure luck.