Restrict \lookup to braces only
This is a prerequisite for the proposal
https://lists.gnu.org/archive/html/lilypond-devel/2023-03/msg00098.html

\lookup used to insert a glyph from either the normal part of the music
font (by default, the emmentaler-<designsize>.otf files) or the brace
part (by default, emmentaler-brace.otf). It does so based on the value
of font-encoding. This is obviously problematic for removing
font-encoding, which is desirable for the reasons explained in the
message linked above.

Note that \lookup is a legacy command: for glyphs from the normal part
of the font, \musicglyph is more typical and does not require
configuring font-encoding; and for braces, \left-brace is more
convenient because it has a well-defined unit.

In the archives of the user list, 100% of uses of \lookup are with
fetaBraces, not fetaMusic.

It is tricky to convert uses for braces to \left-brace because \lookup's
output gets taller with larger font sizes, while it only has an
increased font weight when using \left-brace.

On the other hand, convert-ly can easily handle (non-advanced) uses for
music glyphs (of which there are none in the user list
archives). Therefore, this restricts \lookup to fetaBraces, which should
make it less of a problem going forward while being mostly backwards
compatible (as long as convert-ly is used).