Exit with non-zero return code if Scheme errors were found
This was broken in e12ee410.

Also amend tests that were relying on this using 'expect-error = ##t'
(this is not separate because expect-error complains if no error was
found while one was expected).

Fixes #6344

(cherry picked from commit e76ddc705367cdaa03c0bb28371fceb0473fd730,
changed \version in the regtests to 2.24.2)