Use "normal" instead of "medium" for the regular font weight
In typography speak, the "medium" font weight is an intermediate between
normal and semi-bold.  We want to add this possibility, but this
necessitates changing what LilyPond means by "medium", which until now
meant the normal weight.

This renames the \medium markup command to \normal-weight and the
'medium font-series value to 'normal.