printTrivialVoltaRepeats context property
Setting this option notates `\repeat volta 1` with bar lines and, if
there is an `\alternative`, with a bracket.  The default remains not to
notate it.

The implementation sends volta-span events and sets repeatCommands even
for trivial repeats, but provides repeat/return counts so that engravers
can choose to ignore them based on the value of the new property.

This option is set in InternalGregorianStaff so that a trivial repeat
creates a bar line in a GregorianTranscriptionStaff with the default
settings.  The missing bar line was a shortcoming and is the main
motivation for this feature.

Another potential use is for an ad-lib repeat that the user wants
performed only once in the MIDI.  The user would indicate the
optionality out-of-band, say by using `\alternative` and overriding the
bracket text.