Pass FcMatchPattern to FcConfigSubstitute
As explained by the Fontconfig maintainer in
https://lists.freedesktop.org/archives/fontconfig/2023-April/006970.html
passing FcMatchFont was incorrect. The Fontconfig config allows doing
transformations on font patterns, both before Fontconfig searches for a
font, on the search pattern, and after it has found the font, on the
returned found pattern.  To transform the search pattern, FcMatchPattern
should be passed.

The impact of this change can be seen on

#(ly:message "~s" (ly:font-config-get-font-file "LilyPond Serif"))

which previously returned some default font from the system, and now
correctly returns C059.