SVG backend: Remove broken WOFF font support
This has long been broken; there were already FIXME comments in commit
15eceefc from 2010, and some comments on issue #4648 suggest it never
even worked.  With the Cairo backend being the future of SVG output, I
don't believe this is worth keeping, and it complicates font
enhancements because the special code paths it exercises in the font
code are impossible to test, since WOFF support is broken in the first
place.

Closes #6215, #4648