Restrict Emmentaler font search to LilyPond's data directory
Some Linux distros install Emmentaler in directories found by
Fontconfig.  This is going to be more important in the future when we
turn Emmentaler into a SMuFL font; it will be desirable to enable other
applications to find the fonts like other fonts, not in directories
internal to LilyPond.

This commit introduces a separate Fontconfig configuration object and
corresponding Pango font map that do not know about system font
directories.  These are used when finding Emmentaler in fetaText so that
official binaries of one version of LilyPond unpacked somewhere on the
system will not accidentally pick up Emmentaler files from another
version of LilyPond installed through the distro package.

Fixes #6485