BendSpanner: Use grob-interpret-markup to get text stencil
BendSpanner previously used interpret-markup to interpret markups,
passing (ly:grob-alist-chain grob) for the properties.  Because the
optional argument to ly:grob-alist-chain is not passed, this falls back
to font-defaults, which is intended for music fonts, setting
font-encoding to #'fetaMusic and font-family to #'music.  On top of
this, the BendSpanner grob definition set font-encoding to #'latin1,
resulting in font-encoding = #'latin1 + font-family = #'music, a
combination that doesn't make sense.

Before commit d2d1f3588e, this was handled by falling back to "LilyPond
Serif" due to the configuration in 00-lilypond-fonts.conf.  That commit
disabled font fallback in fetaText, but it does not detect this case
because the font-encoding is latin1.  Because the Fontconfig XML code
was also removed, that changed the fallback to system fonts instead of
LilyPond's default fonts.

Fix this by using grob-interpret-markup, which is the normal interface
to interpret a markup using a grob and reads text-font-defaults rather
than font-defaults.  (This also obviates setting font-encoding in the
BendSpanner definition.)

Fixes #6585