Create InternalMensuralStaff context
Really clean up issue #6586 by creating an InternalMensuralStaff
context and basing MensuralStaff and PetrucciStaff on it.

Fine-tuning the layout is not the purpose of this change, but there are
some noteworthy visual changes to MensuralStaff.

* the start of a repeated section is notated with a single BarLine
* the end of a repeated section is notated with a SignumRepetitionis
* `\fine` creates a thin-thick bar

There are extensive comments explaining why things are the way they are.
This professional software developer is not likely to produce much from
mere hours of amateur musicology, but I hope they will be helpful to
future maintainers.