Doc-ca: Convert translationof to original nodes
As laid out in the mailing list thread on the translated documentation
https://lists.gnu.org/archive/html/lilypond-devel/2023-01/msg00047.html
it is possible to replace the custom setup using translationof with an
approach using the original nodes.

In the new setup, all references (including the menus) are made to the
original English node names, while texi2html and texi2pdf automatically
use the (translated) section names as labels for all internal references
within the same document. For cross-references to other manuals, it is
now required to specify the translated label to display (via the *named
macros), instead of relying on generated xref-maps.

The majority of this change is the result of running the script
convert-translationof.py that is attached to the merge request:
https://gitlab.com/lilypond/lilypond/-/merge_requests/1973
The only manual change in this commit is the details macro in
Documentation/ca/web/manuals.itexi.