Fix \partial in polymeter ossia staff
Add an internal Context method to get the time the Context was
initialized.  Use it in Partial_iterator instead of assuming that the
Timing context was created at moment 0.

There might be other places where this should be used.  Many engravers
use local first-time flags, and some of them might not be modeling what
they should be modeling.  Checking the initialization time of the Timing
context is an idea to explore.

It would probably be useful to expose this in Scheme; however, I don't
want to do that in a stable release without a defect serious enough to
justify it.