Turn `fonts` into a property
Making `fonts` a grob/markup property instead of an output def variable
improves over the existing font-name property in several respects.  For
one thing, it's now possible to write code like

{
  \override DynamicText.fonts.music = "lilyjazz"
  \override DynamicText.fonts.roman = "Noto Serif"
  c'#(make-dynamic-script #{ \markup { p \normal-text dolce } #})
}

which changes the fonts used for "p" and "dolce" independently.

For another, font-name is a Pango font description string, which is a
difficult-to-use interface: it bypasses \bold, \italic and such, instead
requiring to include such modifiers in the string; and it sometimes
requires trailing commas, as with "Times New Roman,".  Values in the
fonts alist are simple font family names and do not have these problems.

The new way to set fonts globally uses the property-defaults variable
introduced in commit ece915a652.