texi2html: Use NODE_FILES
This is meant to "produce one file per node for cross references"
(according to the help of texi2html) and allows to replace the
element_file_name handler by a small handler for node_file_name
that just lower-cases the result of the default handler.

It also produces redirect files for anchors, which we don't rely on
yet but that may be useful in the future. This mode is also turned
on by default in the HTML output of texi2any.