cairo.cc: Don't use `cairo_show_text`
Probably due to a misunderstanding while developing the Cairo backend, glyph
names 'uniXXXX' or 'uXXXXX' were mapped back to Unicode input characters.
However, this didn't work correctly: The code assumed that there was either
a string of Unicode characters derived from those glyph names, or an array
of glyphs where a mapping to Unicode characters was not possible, but not
both at same time.  In reality, however, this actually does happen.  For
example, if an OpenType feature alters a glyph index, it is no longer
mappable to Unicode.

This commit removes the use of a Unicode input string as an input for Cairo.