Documentation: Use headings as appropriate
Chapter 3 on Nodes in the Texinfo manual says: "Normally, you put a
node command immediately before each chapter structuring command" and
at the end of the paragraph "The only time you are likely to use the
chapter structuring commands without also using nodes is if you are
writing a document that contains no cross references and will only be
printed, not transformed into Info, HTML, or other formats."

This certainly does not apply to LilyPond's documentation, which is
produced as PDF and HTML for all languages, and in the Info format
for the English manuals. In fact, the HTML output currently works a
bit by chance because there is custom splitting code in the texi2html
setup code. However, experiments with further simplifications indicate
that it is very easy to run into bugs with this non-standard usage of
commands. For example, texi2html-1.82 will produce numbered files named
as index_NNN.html for structuring commands without associated nodes
(when invoked as texi2html --split=node --node-files --prefix=index).

The appropriate commands for this use case are heading-like commands,
as described in Chapter 4 of the Texinfo manual: They "produce simple
unnumbered headings that do not appear in a table of contents, are not
associated with nodes, and cannot be cross-referenced. These heading
commands never start a new page."