Timing_translator: add Timing alias before init
Add the Timing alias to the context before initializing any translators.
This enables other translators in the same context to find the Timing
context reliably when they are initializing, regardless of the order of
initialization.

This is intended to be simple enough to backport, just in case that
becomes desirable, although the kinds of defects that this feature would
help fix are likely rare.  See ec30a4ffa39ccce8ca7e21d983797f07abf4f384.