Move bar checks to Timing_translator
A comment in the old Bar_check_iterator explained that bar checks were
implemented outside engravers so that they would work in the presence
of skipTypesetting.  That is no longer a concern now that translation
has a pre-process stage that is unaffected by skipTypesetting.

BarCheck music becomes BarCheckEvent.  Timing_translator listens for it
and performs the bar check.  False positive warnings between repeat
alternatives are avoided by ignoring a bar check if either alternative
is aligned at that moment, irrespective of the placement of the `|` in
the musical structure; therefore, false negatives are possible.  This
closes #6602.

The new check uses measureStartNow rather than measurePosition, so it is
not fooled by cadenza material at the beginning of a measure.  This
closes #6550.

Only the first failed bar check in a score (rather, in the context in
which Timing_translator operates) produces a warning.  The previous
implementation suppressed repeated warnings with the same measure
position but would warn again if the offset changed.  That seemed
unnecessarily complicated.

barCheckSynchronize is still supported, maybe.  There were no regression
tests and I haven't tried to add any.  I would like to remove it, but
graciously.