TimeSignature.senza-misura-stencil
Add a `TimeSignature.senza-misura-stencil` property.  The default
`stencil` callback defers to it when `fraction` is `#f` or is not set.

The default is to print nothing, but overriding `senza-misura-stencil`
to `ly:time-signature::print-x` prints an X-shaped sign.

For convenience, property functions `\senzaMisuraTimeSignatureX` and
`\senzaMisuraTimeSignatureOff` change this in the `Timing` context.