postprocess_html: Check all links before appending lang_ext
The check already existed for big-pages, to avoid broken links from
the Downloads page. Enabling the code path unconditionally fixes lots
of broken links in the offline documentation, either to untranslated
nodes (for example the "Long literature list" in the French essay) or
untranslated manuals (the Chinese NR, the IR, and the music glossary).