Move remove_unneeded_anchor to postprocess_html.py
The script is already processing all links, so we might as well do it
there. This has some advantages: First, it allows us to process links
outside of anchors, for example link tags in the header referring to
the next or previous page. Second, we do not process external links
anymore - this is visible by the fact that we are not stripping the
final slash anymore. And finally, it simplifies the texi2html setup
and enables some further work in that area.