Remove validation link from website
Having these links and icons was a fashion some years ago, but that
seems to have mostly faded. And the placement on our website looks
quite ugly (in my opinion). Moreover, HTML 4.01 Transitional is not
the latest and greatest anymore, long live HTML 5!

Funnily enough, the added div actually introduces an error since commit
1edb8b3c85 removed the opening paragraph in the hosting_thanks that is
still closed after the validator link. As far as I can tell, it is the
only validation error...