Beam: clarify a search
A beaming value is a pair of number lists.  ly_memv() was being used to
search for an integer within this pair, i.e., it would first compare the
sought integer to the car of the pair, which could never have matched
because it is a number list, not a number.

Therefore, this call to ly_memv() was effectively searching only the
second list of the pair.  Adding a call to scm_cdr() makes this obvious.