#ifdef DEBUG → if constexpr (CHECKING)
* Rename the preprocessor macro `DEBUG` to `CHECKING` and always define
  it as `0` or `1` to make it usable in `if` statements.
* Update the Contributor's Guide where it mentioned `-DDEBUG`.

With `if constexpr (X)`, the compiler evaluates condition `X` at compile
time and discards one of the alternative statements, but still analyzes
the discarded statement to an extent.  For example, as seen in this
commit, the compiler recognizes that certain variables are used when
checking is enabled, so it does not warn if they are unused when
checking is disabled.

For the checking feature, `if constexpr` increases maintainability, but
it is not a general replacement for `#if`.  For example, it would not
work for a portability condition where the discarded branch called an
undeclared function: that would fail to compile.