css: Fix styling of toc_current
In commit 7c8ea600c3 ("css: Style all ul's in the tocframe"), I replaced
the explicit .toc class selector with the ul type selector. In CSS, type
selectors have a lower specificity which results in some paths with
.toc_current winning over declarations that style nested elements. Solve
this by applying only to the immediate children of .toc_current. This
actually simplifies the reset declarations in lilypond-manuals.css while
lilypond-website.css needs to explicitly reset the black color of the
second-level navigation.