Fix duration-line with partCombine
partCombine distributes the input to several voices. Sometimes there
will be no available ending moment for the duration-line.
Fixed by using ZERO-MOMENT in such cases.

If partCombines prints a chord all note-heads will get a duration-line
now.
This may be a bug or a feature...beyond the scope of this patch, though.