Give `context-mod-from-music` an optional context type argument
This allows refactoring `\settingsFrom` to avoid code duplication and also make
its functionality available as a function defined just in Scheme.

If `\settingsFrom` is intentionally applied to music containing durations,
warnings will now be emitted in contrast to the previous behavior.  Since no
previous uses are actually known, this is deemed acceptable.