binaries: Download zlib from GitHub
The sources for 1.2.13 were deleted from zlib.net after the release
of version 1.3 (which I don't want to update to right now, and this
is not the first time this happens). The download from GitHub seems
to be the more stable.

(cherry picked from commit 7c0f722c059a03eeda2500bb9f6b51c3785ed35f)