Preserve ToC and index links if compiled with new PDF engine of Ghostscript
This is relevant for builds that use the new PDF engine of Ghostscript
before version 10.01, together with `extractpdfmark`, circumventing a bug.

Note that the actual use of the new PDF engine before version 10.02 is not
recommended.  While this commit makes the produced LilyPond documentation
files functional, some output files like the NR are four times larger than
PDFs generated with 10.02 and newer (see
https://lists.gnu.org/archive/html/lilypond-devel/2023-01/msg00103.html for
more details).