Improve rendering of `BassFigureContinuation`
Up to now, the code simply used markup string "0" as an invisible markup
text for the right side of the `BassFigureContinuation` spanner to get its
correct length.  The new code uses digit 0 (which corresponds to input
`\figuremode { <0> }`) and sends it through `figuredBassFormatter`.

As a consequence, the figured bass extender line really ends at the right
edge of a figured bass stack by default, and `BassFigureLine` grobs now have
the same height with and without extender lines – the previously used "0"
markup string resulted in an ordinary digit from a text font, which is
usually much larger than a figured bass digit from the Emmentaler font and
thus vertically enlarged the whole `BassFigureLine` grob.