Figured_bass_engraver::process_music: Fix event handling
If the current moment is within the time span of the last figured bass
event, don't clear spanners.  Without this correction it is possible to get
an incorrect second `BassFigureLine` grob, eventually positioning the
invisible bass figure (which is used to find the continuation line's right
edge) too high.

We already have a regression test for that (`figured-bass-durations.ly`) but
this detail was apparently missed – the placement of the invisible bass
figure can only be seen with activated skyline debugging.

Fixes issues #1513 and #6428.