Implement bar-line markup command
Input is a string which may contain every of "[]{}S|.:!;".
Output is a vertical centered bar line, like the ones for the bar-
command in music. Additionally brace bar lines are coded.
All types scale with fontsize.  Most with height, apart from:
Bracket bar line: The bracket-tips are not scaled with height.
Mimicking SystemStartBracket.
Neither segno bar line, in order to be inline with the behaviour of
the segno bar line in music.
Parameter like hair/thick-thickness and kern are adjustable.
Dashed and dotted have an customizable amount of dashes resp. dots.

Closes #3164