Clef: Use `extra-space` for distances to time and key signatures
Despite of being only three modified lines, this is a huge change, affecting
a lot of scores.

Here is a table of old and new widths as reported by
`ly:paper-column::print` (at 20pt staff size).

| clef + time sig | old  | new  | difference | comment
| --------------- | ---- | ---- | ---------- | ------------------
| `treble`        | 7.70 | 7.58 |  -1.6%     |
| `bass`          | 7.70 | 7.70 |   0%       |
| `alto`          | 7.70 | 7.74 |  +0.5%     |
| `GG`            | 7.70 | 9.08 | +17.9%     | was far too narrow
| `varC`          | 7.70 | 7.27 |  -5.6%     |
| `petrucci-c3`   | 7.70 | 6.02 | -21.8%     | was far too wide

The results for 'clef + key sig' are very similar.

Note that, as before, the widest clef in a staff group determines the
horizontal position of all clefs in a system; this means, for example, that
a piano score containing a treble and a bass clef doesn't change at all.

Fixes issue #3598.