Fix horizontal position of volta brackets not starting at a bar line
Previously, the bar line acknowledger of the `Volta_bracket_engraver` pushed
both the start and the end bar of a volta bracket to the same array, `bars`.
This had the disadvantage that it was not possible to discern a volta
bracket not starting at a bar line from a volta bracket not ending at a bar
line.  Additionally, if there was something in the `bars` array, the code in
function `ly:volta-bracket::calc-shorten-pair` always assumed that it was
the bar line where the volta bracket started.  As a consequence, the
`shorten-pair` values were incorrect for volta brackets not starting at a
bar line.

This commit fixes that; it uses two different arrays `bars-left` and
`bars-right`, also updating the code in
`ly:volta-bracket::calc-shorten-pair` accordingly.