Let Ferneyhough hairpins support al niente circles
Also providing extended regtests and changes.tely

This patch does as the title says, though, it does not deal with
the possibilty of broken hairpins, this is another issue.
Nevertheless, because per default after-line-breaking for Hairpins
is set to ly:spanner::kill-zero-spanned-time, we have to look at the
break status of the Hairpin, if a decrescendo ends at first note of a
new line. We ensure to print a circled tip at the relevant part of the
Hairpin.

Closes #3357