Documentation/GNUmakefile: Fix generation of soft links
The variable `UPDATE_IMAGE_SYMLINKS`, holding a command to create symlinks
for directories, is called multiple times.  Previously, it simply issued

```
ln -sf ../en/[a-f0-9][a-f0-9] .
```

However, it can happen that there aren't such directories yet; as a
consequence, the shell returned the glob as-is, and an (invalid) soft link
called `[a-f0-9][a-f0-9]` got generated.

The new code is recommended in the GNU find info manual.