Direction: Avoid returning & to expiring objects
Qualify member functions with `&` when they return an lvalue reference
to the object or one of its members.  This prevents calling them in some
situations that would create dangling references.

operator*= with an rvalue on the left-hand side had been used in some
tests.  They have been rewritten to use lvalues.