Give glyph->stencil an additional `is-span`-argument
This argument will be set as boolean and used in
bar-line::compound-bar-line resp. span-bar::compound-bar-line.
As a consequence every make-bar-line procedure now needs this
agument in addition to the already mandatory arguments.
The benefit: it is now obvious from inside the make-bar-line
procedure whether some code is for span bar or bar line.