musicxml2ly: Fix handling of ligature brackets
Replace the `print_before_note` and `print_after_note` methods of class
`BracketSpannerEvent` (which are no longer called) with `wait_for_note` and
`print_ly`.

This was broken since LilyPond version 2.19.44 in 2016(!), missed when
integrating some of the 'Philomelos' changes.  A possible reason is that the
`33a-Spanners.xml` regression test was buggy, no longer displaying useful
results then with the updated `musicxml2ly` script.