Improve span bar of dotted bar line
A new printing routine is implemented, distributing
the dots more evenly. While staff-space, taken from layout, is
used to determine the amount of dots (the result is rounded) it is
not strictly used for the vertical gap between staff-line and dot,
resp dot and dot.
The positions of the dots are now computed.

We take staff-space from layout, not from StaffSymbol.staff-space,
because otherwise the topmost staff's space would determine the
whole span bar.