musicxml2ly: Support `<multiple-rest>` element
Multi-measure rests are now only created if the MusicXML input actually
indicates such a rest.

This is a change in behaviour to represent the input more truthfully –
previously, a series of full-measure rests were always converted to a single
multi-measure rest.

It also fixes the appearance of multi-measure rests, listening to the
`use-symbols` attribute.

Also update regression test.