Switch Docker images to Ubuntu 22.04
Even though Ubuntu 20.04 is still supported for about one year, going
to Ubuntu 22.04 has some advantages such as providing newer software
packages: Guile 3.0.7 is a more relevant testing target than the older
Guile 3.0.1 in Ubuntu 20.04, and Texinfo 6.8 allows to drop quite some
compatibility code for Texinfo 6.7 from Documentation/lilypond.init.

The only downside is that, once we require some of the versions found
in Ubuntu 22.04 as a minimum, LilyPond won't build anymore on 20.04.
This is probably fine for master, given that Ubuntu 24.04 was just
released, so we will support the current and previous LTS versions.

Compared to Ubuntu 20.04, explicitly install fonts-dejavu-core and
do not delete the libdl.a and libpthread.a compatibility archives.